backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.21.0",
    "dotenv>=0.9.9",
    "fastapi>=0.109.0",
    "firebase-admin>=6.4.0",
    "mlx-embeddings>=0.0.5",
    "mypy>=1.19.1",
    "numpy>=1.24.0",
    "openai>=1.12.0",
    "pandas>=2.0.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "pypdf>=4.0.0",
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.6",
    "pyyaml>=6.0",
    "qdrant-client>=1.7.0",
    "rich>=13.7.0",
    "scikit-learn>=1.4.0",
    "sqlmodel>=0.0.37",
    "tiktoken>=0.6.0",
    "tqdm>=4.66.0",
    "types-pyyaml>=6.0",
    "uvicorn[standard]>=0.27.0",
    "websockets>=12.0",
]

[[tool.mypy.overrides]]
module = ["mlx_embeddings.*"]
follow_untyped_imports = true
